Make cell harder to mutate or reuse
